Static test suite given by pipeline param
The pipeline now accept a parameter that controls a static test suite.
This is needed in order to merge isv and community pipelines.

In this commit there is also an introduction of common test suite that
contains checks shared between isv and community.

JIRA: ISV-4663
