+++ ansible/roles/operator-pipeline/templates/openshift/pipelines/community-hosted-pipeline.yml
@@ -83,6 +83,10 @@ spec:
       description: The key within the Kubernetes secret that contains the Pipeline SSL key.
       default: operator-pipeline.key
 
+    - name: static_test_suites
+      description: A comma separate list of paths to a test suite that is executed
+      default: operatorcert.static_tests.community,operatorcert.static_tests.common
+
   workspaces:
     - name: repository
     - name: results
@@ -409,8 +413,8 @@ spec:
           value: "$(params.github_token_secret_name)"
         - name: github_token_secret_key
           value: "$(params.github_token_secret_key)"
-        - name: test_suite
-          value: "operatorcert.static_tests.community"
+        - name: test_suites
+          value: "$(params.static_test_suites)"
       workspaces:
         - name: source
           workspace: repository
@@ -634,7 +638,6 @@ spec:
           value: "$(tasks.preflight-trigger.results.artifacts_output_dir)"
       workspaces:
         - name: results
-          subPath: preflight-trigger
           workspace: results
         - name: source
           workspace: repository
