+++ operator-pipeline-images/operatorcert/static_tests/common/__init__.py
@@ -1,0 +1,8 @@
+"""
+A common operators static test suite for all types of operators
+
+A test suite is a python package with two modules called "operator"
+and "bundle" each containing a collection of check functions for the
+respective type of resources.
+
+"""
