+++ operator-pipeline-images/operatorcert/static_tests/isv/bundle.py
@@ -85,22 +85,3 @@ def check_marketplace_annotation(bundle: Bundle) -> Iterator[CheckResult]:
             "To fix this issue define the annotation in "
             "'manifests/*.clusterserviceversion.yaml' file."
         )
-
-
-def check_operator_name(bundle: Bundle) -> Iterator[CheckResult]:
-    """
-    Ensure that the operator name matches the CSV name
-
-    Args:
-        bundle (Bundle): Tested operator bundle
-    """
-    annotation_package = bundle.annotations.get(
-        "operators.operatorframework.io.bundle.package.v1"
-    )
-    if annotation_package != bundle.csv_operator_name:
-        yield Warn(
-            f"Bundle package annotation is set to '{annotation_package}'. "
-            f"Expected value is '{bundle.csv_operator_name}' based on the CSV name. "
-            "To fix this issue define the annotation in "
-            "'metadata/annotations.yaml' file that matches the CSV name."
-        )
