+++ operator-pipeline-images/tests/static_tests/community/test_operator.py
@@ -2,21 +2,20 @@ from pathlib import Path
 from typing import Any
 
 import pytest
-import yaml
 from operator_repo import Repo
 from operator_repo.checks import Fail, Warn
 from operatorcert.static_tests.community.operator import (
     check_ci_upgrade_graph,
-    check_operator_name,
+    check_operator_name_unique,
 )
 from tests.utils import bundle_files, create_files
 
 
-def test_check_operator_name(tmp_path: Path) -> None:
+def test_check_operator_name_unique(tmp_path: Path) -> None:
     create_files(tmp_path, bundle_files("test-operator", "0.0.1"))
     repo = Repo(tmp_path)
     operator = repo.operator("test-operator")
-    assert list(check_operator_name(operator)) == []
+    assert list(check_operator_name_unique(operator)) == []
     create_files(
         tmp_path,
         bundle_files(
@@ -25,7 +24,7 @@ def test_check_operator_name(tmp_path: Path) -> None:
             csv={"metadata": {"name": "other-operator.v0.0.2"}},
         ),
     )
-    assert [x.kind for x in check_operator_name(operator)] == ["failure"]
+    assert [x.kind for x in check_operator_name_unique(operator)] == ["failure"]
 
 
 @pytest.mark.parametrize(
